Added Typescript packages, added basic signup page with routing
