+++ package.json
@@ -7,6 +7,10 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "@types/chess.js": "^0.11.2",
+    "@types/jest": "^27.4.0",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.3",
     "chess.js": "^0.12.1",
     "chessboardjsx": "^2.4.7",
     "react": "^17.0.2",
@@ -14,6 +18,7 @@
     "react-icons": "^4.3.1",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
+    "typescript": "^4.5.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -41,6 +46,6 @@
     ]
   },
   "devDependencies": {
-    "@types/node": "^17.0.15"
+    "@types/node": "^17.0.16"
   }
 }
