+++ src/App.js
@@ -1,24 +1,21 @@
 import React from 'react';
 import './App.css'
-import { Footer } from './containers';
+import { Footer, SignUp } from './containers';
 import { Navbar } from './components';
 import { About, Home } from './views';
 import { Routes, Route } from 'react-router-dom';
-import Matrix from '../src/containers/matrix/index';
+import Matrix from './containers/matrix/index';
 
 export default function App() {
     return (
         <div>
             <div className="App">
-                <div className="gradient_bg">
-                    <Navbar />
-                </div>
                 <Routes>
-                    <Route path="/" element={<Home />} />
-                    <Route path="/about" element={<About />} />
-                    <Route path="*" element={<Matrix/>}/>
+                    <Route path="/" element={<><div className="gradient_bg"><Navbar /></div><Home /><Footer /></>} />
+                    <Route path="/about" element={<><div className="gradient_bg"><Navbar /></div><About /><Footer /></>} />
+                    <Route path="/signup" element={<SignUp/>} />
+                    <Route path="*" element={<Matrix />} />
                 </Routes>
-                <Footer />
             </div>
         </div>
     );
