--- src/views/Home.jsx
@@ -1,10 +1,0 @@
-import { SignUp } from '../containers';
-// import { Brand } from '../components';
-
-const About = () => (
-    <div>
-        <SignUp />
-    </div>
-);
-
-export default About;
