+++ src/views/index.js
@@ -1,2 +1,2 @@
-export { default as About } from './About'
-export { default as Home } from './Home'
+export { default as About } from './About/About'
+export { default as Home } from './Home/Home'
